+++ lib/domain/blocs/account/account_event.dart
@@ -1,3 +1,4 @@
+/*
 part of 'account_bloc.dart';
 
 @freezed
@@ -7,3 +8,4 @@ class AccountDetailsEvent with _$AccountDetailsEvent {
   factory AccountDetailsEvent.loadReset() = AccountDetailsEventLoadReset;
   factory AccountDetailsEvent.logout() = AccountLogoutEvent;
 }
+*/