+++ lib/ui/widgets/account/account_details_cubit.dart
@@ -1,3 +1,5 @@
+/*
+
 // Dart imports:
 import 'dart:async';
 
@@ -55,3 +57,4 @@ class AccountDetailsCubit extends Cubit<AccountDetailsCubitState> {
     await _sessionDataProvider.deleteAccountId();
   }
 }
+*/