+++ lib/ui/widgets/auth/components/header_widget.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 // Project imports:
 import 'package:comics_db_app/ui/components/custom_auth_text_style.dart';
 import 'package:comics_db_app/ui/widgets/auth/components/form_widget.dart';
+import 'package:comics_db_app/ui/widgets/auth/components/guest_auth_button_widget.dart';
 
 class HeaderWidget extends StatelessWidget {
   const HeaderWidget({Key? key}) : super(key: key);
@@ -22,7 +23,14 @@ class HeaderWidget extends StatelessWidget {
             'In order to use the editing and rating capabilities of TMDb, as well as get personal recommendations you will need to login to your account.',
             style: customAuthTextStyle(),
           ),
-          const SizedBox(height: 5),
+          const SizedBox(height: 10.0),
+          Text(
+            'Or you can enter via guest.',
+            style: customAuthTextStyle(),
+          ),
+          const SizedBox(height: 10.0),
+          const GuestAuthButtonWidget(),
+          const SizedBox(width: 5.0),
         ],
       ),
     );